Move API Keep tests working
